Unit 2 lesson 2
Previous commits provided no ability to add a series to our chart.

This commit addes a button to add a series, wires up the click event of
that button to send the `AddSeries` message to our charting actor.

In addition, this commit changes the type of `ChartingActor` from
`UntypedActor` to `ReceiveActor`.