Load project - but with errors
Previous commits actually failed to load in Rider 2018.3.4. (These same
commits loaded in VS 2017 but with compiler errors.) The error reported
seemed to result from trying to load build tools 12.0, but I was unable
to resolve the issue by simply change 12.0 to 15.0 (recommended on
StackOverflow).

To resolve this issue, I:

- Saved all the files from the failing solution directory
- Deleted the original solution directory, `DoThis`
- Created a new F# solution with a console application named `WinTail`
- Copy the solution files back (but not `WinTail.fsproj`)
- Install the NuGet packages:
  - Akka
  - Akka.FSharp
- Add in the existing item, `Actors.fs`

This commit contains the result of those steps.